scx_flow: fix dispatch stall with iterator drain for 4.0.5
Dispatch peeked only the queue head and skipped the move when the head could not run on the asking CPU, so one foreign, exiting, or unresolvable head blocked all later work and stalled the machine under flood.

Drain batch and park with bpf_for_each iterators that skip past bad heads, moving up to 32 tasks per pass with floor advance and deficit accounting per moved task. Park drain shares the same budget so backlog drains in few passes without spinning.

Add Rust drain model plus unit coverage for dead, exiting, and foreign heads and multi-pass drain. Bump to 4.0.5.

+++ scx/src/main.rs
@@ -466,7 +466,7 @@ mod tests {
         );
         assert_eq!(
             crate::flow::LLC_UNKNOWN,
-            crate::bpf_intf::flow_consts_FLOW_LLC_UNKNOWN as u32
+            crate::bpf_intf::flow_consts_FLOW_LLC_UNKNOWN
         );
     }
